Restructure some of the code
