+++ sorting.go
@@ -9,6 +9,32 @@ import (
 	"strings"
 )
 
+// Sort buffer data, for RiveScript.SortReplies()
+type sortBuffer struct {
+	topics map[string][]sortedTriggerEntry // Topic name -> array of triggers
+	thats  map[string][]sortedTriggerEntry
+	sub    []string // Substitutions
+	person []string // Person substitutions
+}
+
+// Holds a sorted trigger and the pointer to that trigger's data
+type sortedTriggerEntry struct {
+	trigger string
+	pointer *astTrigger
+}
+
+// Temporary categorization of triggers while sorting
+type sortTrack struct {
+	atomic map[int][]sortedTriggerEntry // Sort by number of whole words
+	option map[int][]sortedTriggerEntry // Sort optionals by number of words
+	alpha  map[int][]sortedTriggerEntry // Sort alpha wildcards by no. of words
+	number map[int][]sortedTriggerEntry // Sort numeric wildcards by no. of words
+	wild   map[int][]sortedTriggerEntry // Sort wildcards by no. of words
+	pound  []sortedTriggerEntry         // Triggers of just '#'
+	under  []sortedTriggerEntry         // Triggers of just '_'
+	star   []sortedTriggerEntry         // Triggers of just '*'
+}
+
 /*
 SortReplies sorts the reply structures in memory for optimal matching.
